Updated the creation of events, such that they take a template type and default code at definition time.  This code becomes the "EventId" of the event.
We have also changed this templated event to be derived from a BaseEvent
class which has become the pointer type passed around when delivering
events - as you can't know the time of the template specialised derived
class, this is therefore "the" concrete BaseEvent class used in all
references to passing events around the system.

This base event class also now contains, by construction, an instance
copy of the EventId, which we then use for look up of the event during
delivery or event handler callback registry.