Updated the statis create functions, with the hidden constructors std::make_shared was unable to access the class internals, including not being able to get to the private constructor.  We had worked around this with the friendship for _Ref_count_obj, which has worked since about 2011 on MSVC but has broken in VS2019 (v16.x).
Therefore, I have dropped the use of make_sharded and opted to make the
creation functions use the shared_ptr ctr via a call to global operator
new.

And made the global operator new a friend function, and hence granted it
limited access to the internals.

Unfortunately this means that to operator new the ctor is no longer
private, undermining the original intent of this code, however anyone
looking at the header should see the ctor is private and the create
function is public and behave accordingly.